+++ UVtools.WPF/Windows/ToolWindow.axaml.cs
@@ -528,7 +528,6 @@ namespace UVtools.WPF.Windows
                 }
             }
 
-            
 
             // Ensure the description don't stretch window
             DispatcherTimer.Run(() =>